cli: git: extract absolute_git_source() as utility function
